Refactor ActionButton: Remove `displayLabel` and `description` in favor of `labelText`
The `ActionButton` composable has been simplified by removing the `displayLabel` boolean parameter and the `description` string parameter.

A new optional `labelText` string parameter has been introduced. If `labelText` is provided, it will be displayed below the icon and used as the content description for accessibility. If `labelText` is null, no label will be displayed, and the icon will not have a content description.

This change streamlines the API of `ActionButton` and makes the display of the label more explicit.

All call sites of `ActionButton` have been updated to use the new `labelText` parameter.