Implement Switch Profile button
This commit introduces a "Switch to this profile" button in the ProfileSettingsScreen.

The button is displayed when the currently viewed profile is not the active one and there are multiple profiles available.
Clicking the button sets the viewed profile as the active one.

Additionally, the `computeButtonBackground` function in `ActionButton.kt` is now public.